ml-engineer/evidently_audit/requirements.txt
evidently>=0.4.0
scikit-learn>=1.3.0
xgboost>=2.0.0
pandas>=1.5.0
numpy>=1.24.0
pyarrow>=14.0.0
pyyaml>=6.0
